fuse: sanitize lru/invalidate-limit
The default value (64k) for --lru-limit was settled
with change-id Ica2dd4f8f5fde45cb5180d8f02c3d86114ac52b3
(devel branch commit 1778e8e).

The --invalidate-limit option was introduced
with change-id Iddfff1ee2de1466223e6717f7abd4b28ed947788
(devel branch commit 4801fcc). This was, however, an
experimental feature, and this limiting mechanism was
not enabled by default.

Since then, a common wisdom has gathered around
these options, and that proper setting for them
is advised and that it is the right way to handle
certain performance breakdown scenarios.

So at this point there is no point in not applying
the generally beneficial values as defaults.

For --lru-limit, the 64k has proved to be a fine
default. For --invalidate-limit, the usual advice
is then to set it to 130000; but if --lru-limit
needs to be tuned, keep it around the double of it.

We do the following adjustments:
- Set the default value of --lru-limit to 65000, to
  indicate it has no ties with binary arithmetics.
- Besides the direct numerical value, allow
  --invalidate-inode to take a percentage, meaning
  that the limit will take effect when the invalidation
  queue size reaches that percent of lru-limit.
- set 200% to be the default value of --invalidate-limit.
- When a percentage is passed to --invalidate-limit, the
  minimum accepted value is 110%.

Implementation notes:

Although libglusterfs supports not only percent type for
options, but a combined percent or integer type, these were not
suitable for us, because the validator for percentage
enforces a value between 0 and 100, and for our use case,
we need percentages over 100; also the parser just
gives back an unsigned integer with no definite indication
whether it was derived from a percentage or a numeric argument,
and on the call site, it's heuristically categorized by value
(lesser than or equal 100: hey, it's percentage!; bigger than
100: hey, it's a direct number!).

I did not want to touch the existing code which could cause
hard to debug problems if behavior changes in some edge case,
so instead I rolled my own.

I'm introducing the 'percentage_xor_double' type. (The 'xor'
indicates that the parsed value can clearly be identified as
either percentage or (positive) double). With this, there is
no upper bound on the percentage value, either. Percentages
are converted to ratios (ie. value is divided by 100, so that
on the call site they can be just used as multiplier), and
negated; thus the signature is the differentiator for percentage
(ratio) vs. direct number.

Updates: #1000

+++ xlators/mount/fuse/src/fuse-bridge.h
@@ -174,7 +174,7 @@ struct fuse_private {
     /* LRU Limit, if not set, default is 64k for now */
     uint32_t lru_limit;
     uint32_t inode_table_size;
-    uint32_t invalidate_limit;
+    double invalidate_limit;
     uint32_t fuse_dev_eperm_ratelimit_ns;
 
     /* counters for fusdev errnos */
